+++ example/container-middleware.php
@@ -1,10 +1,11 @@
 <?php
 
 use Krak\Mw;
+use Krak\Cargo;
 
 require_once __DIR__ . '/../vendor/autoload.php';
 
-$container = new Pimple\Container();
+$container = Cargo\container();
 $container['i'] = 5;
 $container['inc_mw'] = function() {
     return function($i, $next) {
@@ -15,10 +16,10 @@ $container['inc_mw'] = function() {
 $handler = mw\compose([
     function($i) { return $i; },
     function($i, $next) {
-        $ctx = $next->getContext();
-        return $next($i + $ctx['i']);
+        $c = $next->getContext()->getContainer();
+        return $next($i + $c->get('i'));
     },
     'inc_mw'
-], new Mw\Context\PimpleContext($container));
+], new Mw\Context\ContainerContext($container->toInterop()));
 
 assert($handler(4) == 10);